+++ catalyst/dl/experiment/config.py
@@ -11,11 +11,10 @@ from torch.utils.data import DistributedSampler
 from catalyst.dl.registry import \
     MODELS, CRITERIONS, OPTIMIZERS, SCHEDULERS, CALLBACKS
 from catalyst.dl import utils
-from catalyst.utils.misc import merge_dicts
-from catalyst.utils.hash import get_short_hash
+from catalyst.utils import merge_dicts, get_short_hash, process_model_params
 from catalyst.dl.core import Experiment, Callback
 from catalyst.dl.utils.torch import _Model, _Criterion, _Optimizer, \
-    _Scheduler, process_model_params
+    _Scheduler
 
 
 class ConfigExperiment(Experiment):
