+++ catalyst/dl/utils/__init__.py
@@ -3,8 +3,7 @@
 from .criterion import accuracy, average_accuracy, mean_average_accuracy, \
     dice, f1_score, sigmoid_focal_loss, reduced_focal_loss, iou, jaccard
 
-from .torch import process_components, get_loader, process_model_params, \
-    prepare_cudnn
+from .torch import process_components, get_loader
 # from .trace import *
 from .visualization import plot_metrics
